+++ Source/VisualRecognitionV3/VisualRecognition.swift
@@ -30,7 +30,7 @@ import Foundation
 public class VisualRecognition {
 
     /// The base URL to use when contacting the service.
-    public var serviceURL = "https://gateway-a.watsonplatform.net/visual-recognition/api"
+    public var serviceURL = URL(string: "https://gateway-a.watsonplatform.net/visual-recognition/api")
 
     /// The default HTTP headers for all requests to the service.
     public var defaultHeaders = [String: String]()
@@ -163,10 +163,12 @@ public class VisualRecognition {
         var queryParameters = [URLQueryItem]()
         queryParameters.append(URLQueryItem(name: "version", value: version))
 
+        guard let serviceURL = serviceURL else { return }
+        
         // construct REST request
         let request = RestRequest(
             method: "POST",
-            url: serviceURL + "/v3/classify",
+            url: serviceURL.appendingPathComponent("/v3/classify", isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -226,11 +228,13 @@ public class VisualRecognition {
         // construct query parameters
         var queryParameters = [URLQueryItem]()
         queryParameters.append(URLQueryItem(name: "version", value: version))
+        
+        guard let serviceURL = serviceURL else { return }
 
         // construct REST request
         let request = RestRequest(
             method: "POST",
-            url: serviceURL + "/v3/detect_faces",
+            url: serviceURL.appendingPathComponent("/v3/detect_faces", isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -293,11 +297,13 @@ public class VisualRecognition {
         // construct query parameters
         var queryParameters = [URLQueryItem]()
         queryParameters.append(URLQueryItem(name: "version", value: version))
+        
+        guard let serviceURL = serviceURL else { return }
 
         // construct REST request
         let request = RestRequest(
             method: "POST",
-            url: serviceURL + "/v3/classifiers",
+            url: serviceURL.appendingPathComponent("/v3/classifiers", isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -338,9 +344,10 @@ public class VisualRecognition {
             failure?(RestError.encodingError)
             return
         }
+        guard let serviceURL = serviceURL else { return }
         let request = RestRequest(
             method: "DELETE",
-            url: serviceURL + encodedPath,
+            url: serviceURL.appendingPathComponent(encodedPath, isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -383,9 +390,10 @@ public class VisualRecognition {
             failure?(RestError.encodingError)
             return
         }
+        guard let serviceURL = serviceURL else { return }
         let request = RestRequest(
             method: "GET",
-            url: serviceURL + encodedPath,
+            url: serviceURL.appendingPathComponent(encodedPath, isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -430,11 +438,13 @@ public class VisualRecognition {
             let queryParameter = URLQueryItem(name: "verbose", value: "\(verbose)")
             queryParameters.append(queryParameter)
         }
+        
+        guard let serviceURL = serviceURL else { return }
 
         // construct REST request
         let request = RestRequest(
             method: "GET",
-            url: serviceURL + "/v3/classifiers",
+            url: serviceURL.appendingPathComponent("/v3/classifiers", isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
@@ -508,9 +518,10 @@ public class VisualRecognition {
             failure?(RestError.encodingError)
             return
         }
+        guard let serviceURL = serviceURL else { return }
         let request = RestRequest(
             method: "POST",
-            url: serviceURL + encodedPath,
+            url: serviceURL.appendingPathComponent(encodedPath, isDirectory: false),
             credentials: credentials,
             headerParameters: defaultHeaders,
             acceptType: "application/json",
